refactor: cleanup dead files, confusing names, and trivial re-exports
- Delete src/remote/tui-deploy.ts (2-line re-export), update import in app.tsx
- Inline src/tasks-state.ts into src/tools/registry.ts, update all imports
- Rename src/agent/compact.ts -> llm-summarize.ts and compaction.ts -> artifact-compaction.ts
- Rename exported functions to summarizeMessagesViaLlm and compactMessagesViaArtifacts
- Update all call sites and test file references
- Delete src/util/memory-logger.ts (unused)
