--- src/remote/tui-deploy.ts
@@ -1,1 +1,0 @@
-export { deployForTui, type DeployStep } from "./deploy.js";
