+++ src/ui/task-list.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import { Box, Text } from "ink";
 import Spinner from "ink-spinner";
-import type { Task } from "../tasks-state.js";
+import type { Task } from "../tools/registry.js";
 import { useTheme } from "./theme-context.js";
 import type { Theme } from "./theme.js";
